# src/doc/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"_36_Enums.Role": {
				"type": "string",
				"enum": [
					"COACH",
					"ATHLETE_PROG",
					"ATHLETE_CO",
					"ATHLETE_FULL",
					"ADMIN"
				]
			},
			"Role": {
				"$ref": "#/components/schemas/_36_Enums.Role"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "back",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/user/{id}": {
			"get": {
				"operationId": "FindById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"allOf": [
												{
													"properties": {
														"stat": {
															"items": {
																"properties": {
																	"deadlift": {
																		"type": "number",
																		"format": "double"
																	},
																	"bench": {
																		"type": "number",
																		"format": "double"
																	},
																	"squat": {
																		"type": "number",
																		"format": "double"
																	},
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	}
																},
																"required": [
																	"deadlift",
																	"bench",
																	"squat",
																	"userId",
																	"id"
																],
																"type": "object"
															},
															"type": "array"
														},
														"progUri": {
															"items": {
																"properties": {
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"fileUri": {
																		"type": "string"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"userId",
																	"fileUri",
																	"id",
																	"name"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"stat",
														"progUri"
													],
													"type": "object"
												},
												{
													"properties": {
														"id": {
															"type": "number",
															"format": "double"
														},
														"username": {
															"type": "string"
														},
														"imageUri": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.Role"
														},
														"phone": {
															"type": "string"
														},
														"weight": {
															"type": "number",
															"format": "double"
														},
														"age": {
															"type": "number",
															"format": "double"
														},
														"surname": {
															"type": "string"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"id",
														"username",
														"imageUri",
														"role",
														"phone",
														"weight",
														"age",
														"surname",
														"name"
													],
													"type": "object"
												}
											]
										},
										{
											"type": "string",
											"enum": [
												"NOT-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"allOf": [
												{
													"properties": {
														"stat": {
															"items": {
																"properties": {
																	"deadlift": {
																		"type": "number",
																		"format": "double"
																	},
																	"bench": {
																		"type": "number",
																		"format": "double"
																	},
																	"squat": {
																		"type": "number",
																		"format": "double"
																	},
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	}
																},
																"required": [
																	"deadlift",
																	"bench",
																	"squat",
																	"userId",
																	"id"
																],
																"type": "object"
															},
															"type": "array"
														},
														"progUri": {
															"items": {
																"properties": {
																	"userId": {
																		"type": "number",
																		"format": "double"
																	},
																	"fileUri": {
																		"type": "string"
																	},
																	"id": {
																		"type": "number",
																		"format": "double"
																	},
																	"name": {
																		"type": "string"
																	}
																},
																"required": [
																	"userId",
																	"fileUri",
																	"id",
																	"name"
																],
																"type": "object"
															},
															"type": "array"
														}
													},
													"required": [
														"stat",
														"progUri"
													],
													"type": "object"
												},
												{
													"properties": {
														"id": {
															"type": "number",
															"format": "double"
														},
														"username": {
															"type": "string"
														},
														"imageUri": {
															"type": "string"
														},
														"role": {
															"$ref": "#/components/schemas/_36_Enums.Role"
														},
														"phone": {
															"type": "string"
														},
														"weight": {
															"type": "number",
															"format": "double"
														},
														"age": {
															"type": "number",
															"format": "double"
														},
														"surname": {
															"type": "string"
														},
														"name": {
															"type": "string"
														}
													},
													"required": [
														"id",
														"username",
														"imageUri",
														"role",
														"phone",
														"weight",
														"age",
														"surname",
														"name"
													],
													"type": "object"
												}
											]
										},
										{
											"type": "string",
											"enum": [
												"NOT-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"role": {
										"$ref": "#/components/schemas/Role"
									},
									"password": {
										"type": "string"
									},
									"username": {
										"type": "string"
									},
									"phone": {
										"type": "string"
									},
									"weight": {
										"type": "number",
										"format": "double"
									},
									"age": {
										"type": "number",
										"format": "double"
									},
									"surname": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"role",
									"password",
									"username",
									"phone",
									"weight",
									"age",
									"surname",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"id": {
													"type": "number",
													"format": "double"
												},
												"password": {
													"type": "string"
												},
												"username": {
													"type": "string"
												},
												"imageUri": {
													"type": "string"
												},
												"role": {
													"$ref": "#/components/schemas/_36_Enums.Role"
												},
												"phone": {
													"type": "string"
												},
												"weight": {
													"type": "number",
													"format": "double"
												},
												"age": {
													"type": "number",
													"format": "double"
												},
												"surname": {
													"type": "string"
												},
												"name": {
													"type": "string"
												}
											},
											"required": [
												"id",
												"password",
												"username",
												"imageUri",
												"role",
												"phone",
												"weight",
												"age",
												"surname",
												"name"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"NOT-EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Utilisateur"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/stats": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"deadlift": {
													"type": "number",
													"format": "double"
												},
												"bench": {
													"type": "number",
													"format": "double"
												},
												"squat": {
													"type": "number",
													"format": "double"
												},
												"userId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"deadlift",
												"bench",
												"squat",
												"userId",
												"id"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"USER_NOT_EXIST",
												"ALREADY_EXIST"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Stats"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deadlift": {
										"type": "number",
										"format": "double"
									},
									"bench": {
										"type": "number",
										"format": "double"
									},
									"squat": {
										"type": "number",
										"format": "double"
									},
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"deadlift",
									"bench",
									"squat",
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"deadlift": {
												"type": "number",
												"format": "double"
											},
											"bench": {
												"type": "number",
												"format": "double"
											},
											"squat": {
												"type": "number",
												"format": "double"
											},
											"userId": {
												"type": "number",
												"format": "double"
											},
											"id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"deadlift",
											"bench",
											"squat",
											"userId",
											"id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Stats"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"deadlift": {
													"type": "number",
													"format": "double"
												},
												"bench": {
													"type": "number",
													"format": "double"
												},
												"squat": {
													"type": "number",
													"format": "double"
												},
												"userId": {
													"type": "number",
													"format": "double"
												},
												"id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"deadlift",
												"bench",
												"squat",
												"userId",
												"id"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"STATS_NOT_FOUND"
											]
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Stats"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"deadlift": {
										"type": "number",
										"format": "double"
									},
									"bench": {
										"type": "number",
										"format": "double"
									},
									"squat": {
										"type": "number",
										"format": "double"
									},
									"userId": {
										"type": "number",
										"format": "double"
									}
								},
								"required": [
									"deadlift",
									"bench",
									"squat",
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/stats/{id}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deadlift": {
											"type": "number",
											"format": "double"
										},
										"bench": {
											"type": "number",
											"format": "double"
										},
										"squat": {
											"type": "number",
											"format": "double"
										},
										"userId": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deadlift",
										"bench",
										"squat",
										"userId",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Stats"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deadlift": {
											"type": "number",
											"format": "double"
										},
										"bench": {
											"type": "number",
											"format": "double"
										},
										"squat": {
											"type": "number",
											"format": "double"
										},
										"userId": {
											"type": "number",
											"format": "double"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deadlift",
										"bench",
										"squat",
										"userId",
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Stats"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}